+++ src/components/PieChart.js
@@ -41,7 +41,7 @@ const options = {
           fontsize: 10,
         },
         formatter: function () {
-          return this.key + ": " + this.y + "%";
+          return this.key + ": " + this.y + "/=";
         },
       },
     },
@@ -50,12 +50,12 @@ const options = {
     {
       name: "Share",
       data: [
-        { name: "Planting", y: 61.41 },
-        { name: "Spraying", y: 11.84 },
-        { name: "Weeding", y: 10.85 },
-        { name: "Harvesting", y: 4.67 },
-        { name: "Transportation", y: 4.18 },
-        { name: "Labour", y: 7.05 },
+        { name: "Planting", y: 2000 },
+        { name: "Spraying", y: 1500 },
+        { name: "Weeding", y: 3000 },
+        { name: "Harvesting", y: 1000 },
+        { name: "Transportation", y: 2500 },
+        { name: "Labour", y: 3000 },
       ],
     },
   ],
