+++ src/components/Workers.js
@@ -4,7 +4,7 @@ import { Table } from "react-bootstrap";
 import {  Link } from "react-router-dom";
 //import ExpenditurePieChart from "./ExpenditurePieChart";
 //import SalesBarGraph from "./SalesBarGraph";
-import Navbar from "./Navbar";
+import ProjectAppBar from "./ProjectAppBar";
 //import Reports from "./Reports";
 
 import "../css/index.css";
@@ -35,7 +35,7 @@ class Workers extends Component {
     } else {
       return (
         <>
-          <Navbar />
+          <ProjectAppBar />
           <br></br>
           <div className="container-fluid">
             <div className="row">
