--- src/components/materialDesign/Customers1.js
@@ -1,138 +1,0 @@
-import React from "react";
-import Link from "@material-ui/core/Link";
-import { makeStyles } from "@material-ui/core/styles";
-import Table from "@material-ui/core/Table";
-import TableBody from "@material-ui/core/TableBody";
-import ProjectAppBar from "./ProjectAppBar"
-import TableCell from "@material-ui/core/TableCell";
-import TableHead from "@material-ui/core/TableHead";
-import TableRow from "@material-ui/core/TableRow";
-// Generate Order Data
-function createData(id, name, email, phoneNumber, options, details) {
-  return { id, name, email, phoneNumber, options, details };
-}
-
-const rows = [
-  createData(
-    0,
-    "16 Mar, 2019",
-    "Elvis Presley",
-    "Tupelo, MS",
-    "VISA ⠀•••• 3719",
-    312.44
-  ),
-  createData(
-    1,
-    "16 Mar, 2019",
-    "Paul McCartney",
-    "London, UK",
-    "VISA ⠀•••• 2574",
-    866.99
-  ),
-  createData(
-    2,
-    "16 Mar, 2019",
-    "Tom Scholz",
-    "Boston, MA",
-    "MC ⠀•••• 1253",
-    100.81
-  ),
-  createData(
-    3,
-    "16 Mar, 2019",
-    "Michael Jackson",
-    "Gary, IN",
-    "AMEX ⠀•••• 2000",
-    654.39
-  ),
-  createData(
-    4,
-    "15 Mar, 2019",
-    "Bruce Springsteen",
-    "Long Branch, NJ",
-    "VISA ⠀•••• 5919",
-    212.79
-  ),
-];
-
-function preventDefault(event) {
-  event.preventDefault();
-}
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: "flex",
-  },
-  spacing: {
-    margin: 0,
-  },
-  title: {
-    flexGrow: 1,
-  },
-  grid: {
-    margin: "0px !important",
-  },
-  content: {
-    flexGrow: 1,
-    height: "100vh",
-    overflow: "auto",
-    marginTop: "60px",
-  },
-  container: {
-    paddingTop: theme.spacing(2),
-    paddingBottom: theme.spacing(4),
-    justifyContent: "space-evenly",
-  },
-  paper: {
-    padding: theme.spacing(2),
-    display: "flex",
-    flexDirection: "column",
-  },
-  fixedHeight: {
-    height: 150,
-  },
-}));
-
-export default function Customers1() {
-  const classes = useStyles();
-  return (
-    <div className={classes.root}>
-          <React.Fragment>
-          <ProjectAppBar />
-
-        <main className={classes.content}>
-          <h5 align="left" style={{ marginLeft: "0.5rem" }}>
-            Recent Customers
-          </h5>
-          <Table size="small">
-            <TableHead>
-              <TableRow>
-                <TableCell>Name</TableCell>
-                <TableCell>Email</TableCell>
-                <TableCell>Phone Number</TableCell>
-                <TableCell>Options</TableCell>
-                <TableCell align="right">Details</TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {rows.map((row) => (
-                <TableRow key={row.id}>
-                  <TableCell>{row.name}</TableCell>
-                  <TableCell>{row.email}</TableCell>
-                  <TableCell>{row.phoneNumber}</TableCell>
-                  <TableCell>{row.options}</TableCell>
-                  <TableCell align="right">{row.details}</TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-          <div className={classes.seeMore} style={{ marginLeft: "0.5rem" }}>
-            <Link color="primary" href="#" onClick={preventDefault}>
-              See more orders
-            </Link>
-          </div>
-        </main>
-      </React.Fragment>
-    </div>
-  );
-}
