+++ src/css/customer.css
@@ -12,4 +12,6 @@ width: fit-content;
 .container-fluid1{
 margin-top: -55px;
 }
-
+.table-headings{
+color: white;
+}
